+++ dotnet/resources/NeptuneEvo/Character/Save/Repository.cs
@@ -55,7 +55,7 @@ namespace NeptuneEvo.Character.Save
                     .Set(v => v.Armor, characterData.Armor)
                     .Set(v => v.Lvl, characterData.LVL)
                     .Set(v => v.Exp, characterData.EXP)
-                    .Set(v => v.Money, Convert.ToInt32(characterData.Money))
+                    .Set(v => v.Money, Convert.ToInt64(characterData.Money))
                     .Set(v => v.Bank, characterData.Bank)
                     .Set(v => v.Work, characterData.WorkID)
                     .Set(v => v.Drugaddi, characterData.DrugsAddiction)
